Completion of terminology
Noticed that 'List of Abbreviations' and 'List of Stuctures' in the atlas contain different number of brain regions and do not contain the same brain regions (some are only in one but not the other and some are in both).
The two lists were combined. Updated terminology contains 952 PEVs instead of 809 PEVs from before.